+++ subprojects/zap-clientapi/src/main/java/org/zaproxy/clientapi/core/AlertsFile.java
@@ -20,8 +20,9 @@
 package org.zaproxy.clientapi.core;
 
 import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.List;
 import org.jdom.Document;
@@ -77,8 +78,8 @@ public class AlertsFile {
         XMLOutputter xmlOutput = new XMLOutputter();
 
         xmlOutput.setFormat(Format.getPrettyFormat());
-        try {
-            xmlOutput.output(doc, new FileWriter(outputFile));
+        try (OutputStream os = Files.newOutputStream(outputFile.toPath())) {
+            xmlOutput.output(doc, os);
             System.out.println("alert xml report saved to: " + outputFile.getAbsolutePath());
         } catch (IOException e) {
             e.printStackTrace();