+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Disable XXE processing when parsing ZAP API responses.
+- Ensure alerts file is always closed.
 
 ## [1.6.0] - 2018-04-10
 ### Added